Disable minimap in monaco editor
It seems to me unlikely that our challenges, as quick and bite-sized as
they are, will be long enough to make a minimap useful so it's just
taking up space. Of course i might be wrong about this, so later on we
can either add it back or make it user configurable.