Move prettier formatting into worker
This will likey provide a better experience for the user since they no
longer have to download all of prettier with the inital bundle. Prettier
is _HUGE_.

But more importantly this was breaking the build, taking up too much
memory in Node to complete without setting --max-old-space-size=4096 to
bump node max allowed memory.