Bring back codepress API for data fetching
We were (and still are) in a state of rapid changes. Today we had to get
the server up and running asap so that a student could use it. I've just
added back in the codepress data fetching if we're in codepress mode.
The overall code still needs to be refactored once we've settled our
deployment strategy a bit.