Fix expectation errors
Part of the issue was that the test itself was failing. I forgot to
write `test` anywhere in the test code and just ran expectations
directly. This is not great, but it's even worse that there isn't
currently a good way to surface that information to us, the codepress
user.