Place Button in Link, rather than the other way around
When its the other way around what happens is the link doesn't fill the
button, so its possible to click the button and have nothing happen. In
reality it's just that you didn't click the actual <a> tag within the
button, but in practice it's super unintuitive and very poor UX. Fixed
now.