Import headingFromSlug directly from rich-markdown-editor, smooth scroll to headings
The issue with headingFromSlug was that our version was better! However,
rich-markdown-editor is currently in charge of rendering the Headings
and thus they use their own headingToSlug. Their version will output
"id.." rather than "id", i.e. it will not strip characters such as "."
which may be invalid in some circumstances, such as when used in a
querySelector call.

This also adds smooth scroll support with a fallback that defaults to
default browser behavior.