Default blob loading to true
If the idea is that changing the code in the monaco editor is done via
initializing with the proper code, then we have to ensure that the class
re-initializes every time the code needs to change. Edit mode updates,
such as starterCode/solutonCode are a possible exception, but otherwise
we need re-initialization. This in turn means that we need to ensure our
loading states are on point, otherwise even if data is fetched the
component will not update its contents.