Support searching challenges in the UI (#71)
* Add lunr as dependency

* Add a tmp dir ignore so that there is an area for scratch paper

* Add new constants and types for search worker

* Separate out window dimensions from constants

The issue was that winow is not available within workers so we could not
use the constants file in any env that was not the main browser.

* Add a search worker to do full text search

* Add the Worker type to code worker

* Add search functionality to challenges module

* Add debugger

* Add search input

* Add some search match mapping logic

* Add rudimentary search display logic

* Improve search results styling

* More styling improvements

* Improve code comments and typing in search worker

* Add score to search result type

Not sure that we will need this. Results are already sorted by Lunr for
us. But just in case I added it in.

* Fix zindex issue with search box and move search title box to the bottom

* Further refinements to search box styling

* Use match rather than stemmed term for match context

* Update search match context typing

* Refactor match context logic

* Use auto-css wrapping on returned search matches

* Update excerpt logic to provide a more aesthetic experience

* Reorder search box file

* Flat map!

* Fix spelling on search field!

* Add react-string-replace

* Add click outside logic, title highlighting, z-index fixes to search

* Fix search box scroll styling

* Ignore vscode settings

* Improve formatting of search excerpts by stripping markdown

* Fix comment spelling

* Update linting rules to allow leading underscore

* Update search excerpt padding to put matches more towards the left

* Minor formatting update to challenge

* Support mapping more individual keys

Needed to map the arrow keys and enter

* Add keyboard shortcuts to search menu

* Fix styling of challenge type selector in sandbox following search bar addition

* Rename mapState to mapStateToProps

* Implement link-search in the content editor

Should not affect our users since they use they readonly editor for
displaying content, but for our own purposes this should make it quite
easy to link between pages.

* Update sandbox mapping

I updated this mapping because cmd+k is the mapping used in the editor
for creating a link. I think this is semi-standard as Evernote and some
other editors do the same. Anyway, we don't control that underlying
component so we can't updated it's hotkey.

Co-authored-by: Sean Smith <sean.smith.2009@gmail.com>