Fully hide the editing controls when requested
This allows the header to look just as it would to a user. By using
display none we let the rest of the UI flow in around it. This is
different than jsut updating opacity because now we have the search bar
that is also competing for the same screen space.