Just some misc styling tweaks (#76)
* Tighted up header styling in codepress

The header was getting really tight with the search bar now

* Fully hide the editing controls when requested

This allows the header to look just as it would to a user. By using
display none we let the rest of the UI flow in around it. This is
different than jsut updating opacity because now we have the search bar
that is also competing for the same screen space.

* Improve search result box styling on small screens

* Hide account button on mobile menu if not logged in

* Fix content header not being able to wrap

* Fix TOC not scrolling on smaller screens

* Update mobile view copy