Update course shape
* content -> instructions
* supplementaryContent -> content

What was formerly the "content" is actually meant to be extremely
concise, even terse, and the purpose was to tell the user what they had
to do to pass the challenge. Thus "instructions" seems like a better
name.

What was formerly "supplementary content" is meant to be our long-form
explanations, additional information and indeed supplementary content.
However, for many challenges it is not supplementary but core to the
challenge and the name "content" is no longer taken. Thus the renaming.