Improve markdown shortcuts, solve space key issue (#80)
* Update course shape

* content -> instructions
* supplementaryContent -> content

What was formerly the "content" is actually meant to be extremely
concise, even terse, and the purpose was to tell the user what they had
to do to pass the challenge. Thus "instructions" seems like a better
name.

What was formerly "supplementary content" is meant to be our long-form
explanations, additional information and indeed supplementary content.
However, for many challenges it is not supplementary but core to the
challenge and the name "content" is no longer taken. Thus the renaming.

* Further refinements to table of contents

* Fix line height of content headers

* Wrap table of content link text

* Remove unused challenge generator

This file was used before we had codepress. No longer needed.

* Lock slate types

* Add augmented onKeydown handler to make editing more intuitive

* Fix markdown parsing errors related to spacing by replacing the onSpace handler

* Add in new inline shortcuts

* Fix issue with only checkboxes for list items

* Fix issue with single-character code snippets and code blocks