Update Torvalds to test for challenges without tests
Any challenge with a workspace is assumed to have a test suite. This
assumption could be relaxed at some later point but for now the
reasoning is: If you have a workspace, you want to test some user code.
If you are just instructing, then use a challenge type of
media/section/project/etc where there is no assumption of test code.