Fix delayed prop updates for font-size and theme
If you clicked change theme, the last them (not the one you want) would
be set. Since it's a toggle this meant that it worked fine after the
first click. But that first click appeared to do nothing.

The font size was the same. If you clicked font-up twice, then font-down
once, the font wouldn't go down becuase it was loading the font-size
from prevProps instead of this.props (current props).