Add workable success modal state
This makes it actually usable. Now the modal will only show if the user
pressed the run button. So we can run tests all we want via other parts
of the code calling the run method, but the modal will only show on
success if the user clicks the button.