Add extra monaco libs to JS libs as well as TS
Currently this is mainly for our test editor which is a JS editor rather
than TS.

On that note, the reason for keeping JS is that as far as I can tell TS
doesn't gain us anything in testing. It could potentially help us avoid
type and syntax errors but we are often testing code that is supposed to
do odd thing and we're writing tests, not app code, which we immediately
check for feedback.