Add generated metadata.json file
This might be best ignored, i'm not yet sure. It's a generated file
living on our src directory which feels odd. It should be apparent to
anyone viewing the file that we would not manually update a 12k line
JSON file every time we change the course, but it's still odd.