Fix lint error
Just went ahead and ignore it because it was disagreeing with prettier
and it seemed to be disagreeing with other parts of the codebase. I seem
to remember the rule being no quotes unless needed on a prop by prop
basis, rather than consistency for all props (I.e. if one prop has
quotes then the rest must also)