Update git metadata script to better-account for initial commits to a challenge
The trick here is to combine our knowledge of our own data with gits
code-copy detection. We were already using git to detect lines simply
moved rather than created, but now we also treat the line that commit
the "id" of the challenge as the initial commit of that challenge. This
assumption should hold for all our challenges since ids do not change.

Gits copy detection was finding lines such as "videoUrl: ''" and
thinking the line was copied becuase it is indeed identical to other
lines in the file. This is wrong since it's just a default value, not a
copy value. Git treating it as a copy mean that the earliest commit
detected for a challenge was usually very early indeed, thinking that
this line had been copied from elsewhere.