Reinstate root redirect
It's removal was unintentional. It's a catchall which means we don't
have a 404 possibility, which I would actually support. If something
doesn't exist its fine that the user knows that unless it's something
that moved in which case it definitely should redirect. Anyway, just
reinstating what we had before here.