Remove chat widget
So, what the hell? Did we really dislike the chat widget? Was it causing
user confusion??

Nope. Nope.

7 days ago we added this widget. 7 days ago google shut down our ad
citing "malicious software". Clearly, they are wrong. In fact, even that
is too much an assumption. I don't _know_ if there's a causal
relationship between adding papercups and google shutting down our ad.
However, I imagine that google in the past has faced many a bad actor
who put up something totally legitimate at first in order to gain
approval and then changed it later. It stands to reason that google
would be doing checkups on your ad endpoint to see if things changed,
and if so what.

Still, what the hell? I would love more of an explanation. Why would
they mark papercups as malicious? If it's not because of paper cups then
what? To my knowledge, nothing else has changed other than content. No
more external libs that before.

We don't know what the issue is yet but it would be good to get our site
un-un-approved so I'm just removing this.